fix(editor): 调整编辑器 toolbar action 显示位置
将以下 action 固定在 toolbar 前台显示 (SHOW_AS_ACTION_ALWAYS):
- EditorEditLineMenuAction
- RedoAction
- UndoAction

将以下 action 收纳到 overflow menu (SHOW_AS_ACTION_NEVER):
- FindActionMenu
- ProjectSyncAction
- ReloadColorSchemesAction
- FindInPathAction (已从 EditorActivityActions.kt 迁移到 FindActionMenu 子项)

迁移 FindInPathAction 从 actions/search/ 包到 actions/etc/ 包, 与 FindInFileAction / FindInProjectAction 同包.
