+++ core/app/src/main/java/com/itsaky/androidide/actions/etc/ReloadColorSchemesAction.kt
@@ -18,6 +18,7 @@
 package com.itsaky.androidide.actions.etc
 
 import android.content.Context
+import android.view.MenuItem
 import androidx.core.content.ContextCompat
 import androidx.lifecycle.lifecycleScope
 import com.itsaky.androidide.R
@@ -58,4 +59,10 @@ class ReloadColorSchemesAction(context: Context, override val order: Int) : Edit
     }
     return true
   }
+
+  /**
+   * Collapse this action into the toolbar overflow menu; it should not occupy a fixed toolbar
+   * slot.
+   */
+  override fun getShowAsActionFlags(data: ActionData): Int = MenuItem.SHOW_AS_ACTION_NEVER
 }
