+++ core/app/src/main/java/com/itsaky/androidide/actions/menu/EditorEditLineMenuAction.kt
@@ -1,8 +1,8 @@
-// by android_zero
 package com.itsaky.androidide.actions.menu
 
 import android.content.Context
 import android.util.Log
+import android.view.MenuItem
 import androidx.core.content.ContextCompat
 import com.itsaky.androidide.actions.*
 import com.itsaky.androidide.actions.editor.cursor.CursorNextLocationAction
@@ -109,4 +109,11 @@ class EditorEditLineMenuAction(context: Context, override val order: Int) :
     )
     return true
   }
+
+  /**
+   * Always show this action in the editor toolbar, similar to other fixed toolbar entries like
+   * [com.itsaky.androidide.actions.build.QuickRunWithCancellationAction] and
+   * [com.itsaky.androidide.actions.build.RunTasksAction].
+   */
+  override fun getShowAsActionFlags(data: ActionData): Int = MenuItem.SHOW_AS_ACTION_ALWAYS
 }
